Final Sprint 1 Commit
Here we implemented backend API with Flask for managing investors, stocks, bonds, and transactions. Set up MySQL database integration for CRUD operations on investors, stocks, bonds, and stock/bond transactions. Configured RESTful endpoints using Flask Blueprints for scalability. All of the code is in this one file, and this is the last commit submission.